Edting cabin

+++ src/services/apiCabins.js
@@ -9,38 +9,54 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
-    "/",
-    ""
-  );
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.toString().startsWith(supabaseUrl);
+  const imageName = hasImagePath
+    ? ""
+    : `${Math.random()}-${newCabin.image.name}`.replaceAll("/", "");
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
   // https://zdvgtqrfmtgutdqjrscb.supabase.co/storage/v1/object/public/cabin-images/cabin-002.jpg
 
-  // //1. Create Cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  //1. Create/Edit Cabin
+  let query = supabase.from("cabins");
+  // const { data, error } = await supabase
+  //   .from("cabins")
+  //   .insert([{ ...newCabin, image: imagePath }])
+  //   .select()
+  //   .single();
+
+  //a)Create
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  //b)Edit
+  if (id)
+    query = query.update([{ ...newCabin, image: imagePath }]).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
     throw new Error("Cabin could not be Created");
   }
 
-  //2. Uploading Image
-  const { error: storageError } = await supabase.storage
-    .from("cabin-images")
-    .upload(imageName, newCabin.image);
-  //3. Deleting the cabin if there was a error uploading image
-  if (storageError) {
-    await supabase.from("cabins").delete().eq("id", data.id);
-    console.error(storageError);
-    throw new Error(
-      "Cabin image could not be uploaded and the cabin was not created"
-    );
+  if (!hasImagePath) {
+    //2. Uploading Image
+    const { error: storageError } = await supabase.storage
+      .from("cabin-images")
+      .upload(imageName, newCabin.image);
+    //3. Deleting the cabin if there was a error uploading image
+    if (storageError) {
+      await supabase.from("cabins").delete().eq("id", data.id);
+      console.error(storageError);
+      throw new Error(
+        "Cabin image could not be uploaded and the cabin was not created"
+      );
+    }
   }
+
   return data;
 }
